+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - rendering DC20 feats
 - DC20 class feats
 - rendering Daggerheart weapon distances in squares
+- selecting character path for DC20 characters
 
 ### Modified
 - refreshing character after domain cards and equipment changes